Windows: shell out to tzinfo in a way that's safe for Go 1.18 and lower
Until Go 1.19, the `os/exec` package from standard library had a behavior of implicitly including the current directory when looking up commands to execute.

That means that for every Windows Go program that imports and invokes this library, Go could execute a malicious `./tzinfo.exe` from the current directory.

Go 1.19 fixes this beavior, but Go 1.18 and older are vulnerable. This fix makes Go 1.18 and older versions use the `execabs` package that is designed to patch the security risk.